refactor waypoint logic
