+++ pilot/pkg/networking/core/v1alpha3/listener_waypoint.go
@@ -49,51 +49,10 @@ import (
 	xdsfilters "istio.io/istio/pilot/pkg/xds/filters"
 	"istio.io/istio/pkg/config"
 	"istio.io/istio/pkg/config/host"
-	"istio.io/istio/pkg/config/labels"
 	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/proto"
 	"istio.io/pkg/log"
 )
-
-const (
-	// ConnectTerminate is the name for the resources associated with the termination of HTTP CONNECT.
-	ConnectTerminate = "connect_terminate"
-
-	// MainInternalName is the name for the resources associated with the main (non-tunnel) internal listener.
-	MainInternalName = "main_internal"
-
-	// ConnectOriginate is the name for the resources associated with the origination of HTTP CONNECT.
-	ConnectOriginate = "connect_originate"
-)
-
-type WorkloadAndServices struct {
-	WorkloadInfo *model.WorkloadInfo
-	Services     []*model.Service
-}
-
-func FindAssociatedResources(node *model.Proxy, push *model.PushContext) ([]WorkloadAndServices, map[host.Name]*model.Service) {
-	wls := []WorkloadAndServices{}
-	scope := node.WaypointScope()
-	workloads := push.WorkloadsForWaypoint(scope)
-	for _, wl := range workloads {
-		wls = append(wls, WorkloadAndServices{WorkloadInfo: wl})
-	}
-	svcs := map[host.Name]*model.Service{}
-	for i, wl := range wls {
-		for _, ns := range push.ServiceIndex.HostnameAndNamespace {
-			svc := ns[wl.WorkloadInfo.Namespace]
-			if svc == nil {
-				continue
-			}
-			if labels.Instance(svc.Attributes.LabelSelectors).SubsetOf(wl.WorkloadInfo.Labels) {
-				svcs[svc.Hostname] = svc
-				wl.Services = append(wl.Services, svc)
-			}
-		}
-		wls[i] = wl
-	}
-	return wls, svcs
-}
 
 func (lb *ListenerBuilder) serviceForHostname(name host.Name) *model.Service {
 	return lb.push.ServiceForHostname(lb.node, name)
@@ -105,7 +64,7 @@ func (lb *ListenerBuilder) buildWaypointInbound() []*listener.Listener {
 	// 1. Decapsulation CONNECT listener.
 	// 2. IP dispatch listener, handling both VIPs and direct pod IPs.
 	// 3. Encapsulation CONNECT listener, originating the tunnel
-	wls, svcs := FindAssociatedResources(lb.node, lb.push)
+	wls, svcs := findAssociatedResources(lb.node, lb.push)
 
 	listeners = append(listeners,
 		lb.buildWaypointInboundConnectTerminate(),
@@ -221,7 +180,6 @@ func (lb *ListenerBuilder) buildWaypointInternal(wls []WorkloadAndServices, svcs
 			portString := fmt.Sprintf("%d", port.Port)
 			cc := inboundChainConfig{
 				clusterName: model.BuildSubsetKey(model.TrafficDirectionInboundVIP, "tcp", svc.Hostname, port.Port),
-				// clusterName: model.BuildSubsetKey(model.TrafficDirectionInboundPod, "", host.Name(wl.PodIP), port.Port),
 				port: ServiceInstancePort{
 					Name:       port.Name,
 					Port:       uint32(port.Port),
@@ -275,7 +233,7 @@ func (lb *ListenerBuilder) buildWaypointInternal(wls []WorkloadAndServices, svcs
 	{
 		// Direct pod access chain.
 		cc := inboundChainConfig{
-			clusterName: EncapCluster.Name,
+			clusterName: EncapClusterName,
 			port: ServiceInstancePort{
 				Name:     "unknown",
 				Protocol: protocol.TCP,
@@ -699,7 +657,7 @@ func (lb *ListenerBuilder) GetDestinationCluster(destination *networking.Destina
 	}
 
 	if service != nil {
-		_, svcs := FindAssociatedResources(lb.node, lb.push)
+		_, svcs := findAssociatedResources(lb.node, lb.push)
 		_, f := svcs[service.Hostname]
 		if !f || service.MeshExternal {
 			// this waypoint proxy isn't responsible for this service so we use outbound; TODO quicker lookup
